Fundamentally changed the architecture of the labsight library. Instead the library returning motor objects for connected motors, the library now simply returns a dictionary of ids and serials. This means is that the gui is now responsible for creating motor objects and assigning serials to them. The reason for this is that the gui can now create motor objects with an empty serial and connect to the serial later. This is useful for guis because it can show motors that have previously been configured but don't happen to be plugged in at the moment.
Also made quite a bit of progress on the gui. The gui scans configuration files for motors and allows the configuration of motors that have not yet been configured. If the motor is connected during this scan, it will connect to the serial port for this motor. Later, I will make it so that the gui can connect to motor serials even after this initial scan for configuration files.